backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10"
aiogram = "^3.17.0"
fastapi = "^0.115.6"
tortoise-orm = "^0.23.0"
pydantic-settings = "^2.7.1"
uvicorn = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
